refactor: modularize HeroMockups component into a structured directory of scenes and sub-components
